@@ -1,7 +1,7 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk">
 
   <PropertyGroup>
-   <TargetFrameworks>netstandard2.0;net6.0;net7.0;net8.0</TargetFrameworks>
+   <TargetFrameworks>netstandard2.0;net8.0;net9.0</TargetFrameworks>
     <Description>Entity Framework Core SQL provider-neutral job storage implementation for Hangfire (https://www.hangfire.io).</Description>
     <PackageReleaseNotes>
 0.6.0
@@ -54,19 +54,14 @@ This is the first public release.
   </ItemGroup>
 
   <Choose>
+    <When Condition=" '$(TargetFramework)'=='net9.0' ">
+      <ItemGroup>
+        <PackageReference Include="Microsoft.EntityFrameworkCore.Relational" Version="9.0.0" />
+      </ItemGroup>
+    </When>
     <When Condition=" '$(TargetFramework)'=='net8.0' ">
       <ItemGroup>
         <PackageReference Include="Microsoft.EntityFrameworkCore.Relational" Version="8.0.0" />
-      </ItemGroup>
-    </When>
-    <When Condition=" '$(TargetFramework)'=='net7.0' ">
-      <ItemGroup>
-        <PackageReference Include="Microsoft.EntityFrameworkCore.Relational" Version="7.0.0" />
-      </ItemGroup>
-    </When>
-    <When Condition=" '$(TargetFramework)'=='net6.0' ">
-      <ItemGroup>
-        <PackageReference Include="Microsoft.EntityFrameworkCore.Relational" Version="6.0.0" />
       </ItemGroup>
     </When>
     <Otherwise>
