@@ -1,26 +1,20 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk">
 
   <PropertyGroup>
-    <TargetFrameworks>net48;net6.0;net7.0;net8.0</TargetFrameworks>
+    <TargetFrameworks>net48;net8.0;net9.0</TargetFrameworks>
   </PropertyGroup>
 
   <Choose>
+    <When Condition=" '$(TargetFramework)'=='net9.0' ">
+      <ItemGroup>
+        <PackageReference Include="Microsoft.EntityFrameworkCore.Sqlite" Version="9.0.0" />
+        <PackageReference Include="Microsoft.Extensions.Logging.Debug" Version="9.0.0" />
+      </ItemGroup>
+    </When>
     <When Condition=" '$(TargetFramework)'=='net8.0' ">
       <ItemGroup>
         <PackageReference Include="Microsoft.EntityFrameworkCore.Sqlite" Version="8.0.2" />
         <PackageReference Include="Microsoft.Extensions.Logging.Debug" Version="8.0.0" />
-      </ItemGroup>
-    </When>
-    <When Condition=" '$(TargetFramework)'=='net7.0' ">
-      <ItemGroup>
-        <PackageReference Include="Microsoft.EntityFrameworkCore.Sqlite" Version="7.0.16" />
-        <PackageReference Include="Microsoft.Extensions.Logging.Debug" Version="7.0.0" />
-      </ItemGroup>
-    </When>
-    <When Condition=" '$(TargetFramework)'=='net6.0' ">
-      <ItemGroup>
-        <PackageReference Include="Microsoft.EntityFrameworkCore.Sqlite" Version="6.0.27" />
-        <PackageReference Include="Microsoft.Extensions.Logging.Debug" Version="6.0.0" />
       </ItemGroup>
     </When>
     <Otherwise>
