@@ -6,9 +6,12 @@ internal static class StateExtensions
 {
     internal static DateTime? GetCreatedAt(this IState state)
     {
+#if NET8_0_OR_GREATER
+        ArgumentNullException.ThrowIfNull(state);
+#else
         if (state is null)
             throw new ArgumentNullException(nameof(state));
-
+#endif
         return state switch
         {
             ScheduledState scheduled => scheduled.ScheduledAt,
